% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgDistort.R
\name{.calcDelta}
\alias{.calcDelta}
\title{Calculate delta between averages of two years.}
\usage{
.calcDelta(
  data,
  weightCol,
  yearCol,
  year0,
  year1,
  variables,
  indexNames,
  outRow = NULL,
  y0Value = NULL,
  type = c("absolute", "relative")
)
}
\arguments{
\item{data}{A data.frame containing all columns specified in \code{weightCol}, \code{yearCol}, \code{indexList}, and \code{variables}.}

\item{weightCol}{The column in \code{data} that contains the weights.}

\item{yearCol}{The column in \code{data} that contains the year.}

\item{year0}{The reference year to which the delta should be calculated.}

\item{year1}{The subsequent year.}

\item{variables}{The variables for which the calculation should be executed. Consult the example in \code{\link{findInfluencialObservations}} to see how this argument must be structured.}

\item{indexNames}{The names of the columns that contain the indices.}

\item{outRow}{The special row for which the index level should be filtered.}

\item{y0Value}{Optional precalculated value for the reference year, in case the calculations should be carried out many times (to speed up the process).}

\item{type}{The type of delta. If \code{"relative"}, then the relative difference will be calculated as: \code{(year1 / year0) - 1}}
}
\value{
The delta for all variable and index combinations. Can be a matrix or a vector depending on the inputs.
}
\description{
Calculate delta between averages of two years.
}
\author{
Daniel Hoop
}
\keyword{internal}
