% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgDistort.R
\name{.calcYear1}
\alias{.calcYear1}
\title{Calculate average of one year}
\usage{
.calcYear1(
  data,
  weightCol,
  yearCol,
  year,
  variables,
  indexNames,
  outRow = NULL
)
}
\arguments{
\item{data}{A data.frame containing all columns specified in \code{weightCol}, \code{yearCol}, \code{indexList}, and \code{variables}.}

\item{weightCol}{The column in \code{data} that contains the weights.}

\item{yearCol}{The column in \code{data} that contains the year.}

\item{year}{The year to calculate the average for.}

\item{variables}{The variables for which the calculation should be executed. Consult the example in \code{\link{findInfluencialObservations}} to see how this argument must be structured.}

\item{indexNames}{The names of the columns that contain the indices.}

\item{outRow}{The special row for which the index level should be filtered.}
}
\value{
The average values for all variable and index combinations. Can be a matrix or a vector depending on the inputs.
}
\description{
Calculate average of one year
}
\author{
Daniel Hoop
}
\keyword{internal}
