% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgDistort.R
\name{.condenseResults}
\alias{.condenseResults}
\title{Condense results.}
\usage{
.condenseResults(
  x,
  indexList,
  variables = NULL,
  weightResult = FALSE,
  giveNames = FALSE
)
}
\arguments{
\item{x}{The data structure containing the results to be condensed. A return value of the function \code{\link{influenceOfGivenObservation}}.}

\item{indexList}{A list which holds all index levels which should be checked. Consult the example in \code{\link{findInfluencialObservations}} to see how this argument must be structured.}

\item{variables}{The variables for which the calculation should be executed. Consult the example in \code{\link{findInfluencialObservations}} to see how this argument must be structured.}

\item{giveNames}{Logical indicating if the result should be given rownames equal to the index levels.}

\item{weightResults}{Logical value indicating if the influence of the observation on the different variables should be weighted.}
}
\description{
Condense results such that indexes levels that aren't relevant for an observation aren't displayed. If variables are given, then they will be weighted in order to further condense.
}
\author{
Daniel Hoop
}
\keyword{internal}
