% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgDistort.R
\name{.createRowFinderFunc}
\alias{.createRowFinderFunc}
\title{Define function to get index of observations.}
\usage{
.createRowFinderFunc(missingIn, data, idCol, yearCol, year0, year1)
}
\arguments{
\item{missingIn}{A character defining the method to apply. Either "anyYear", "year0" or "year1".}

\item{data}{A data.frame containing all columns specified in \code{weightCol}, \code{yearCol}, \code{indexList}, and \code{variables}.}

\item{idCol}{Optional: The name of the column in \code{data} which contains the ids. Must be specified if \code{ids} is not \code{NULL}.}

\item{yearCol}{The column in \code{data} that contains the year.}

\item{year0}{Optional: The reference year to which the delta should be calculated. Only has to be given if \code{method = "oneYear"} is chosen.}

\item{year1}{Optional: The subsequent year. Only has to be given if \code{method = "oneYear"} is chosen.}
}
\value{
A function that will can be used to perform the task inside a call to \code{\link{tapply}}.
}
\description{
Define function to get index of observations.
}
\author{
Daniel Hoop
}
\keyword{internal}
