% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgDistort.R
\name{influenceOfGivenObservation}
\alias{influenceOfGivenObservation}
\title{Calculate the influence of a given observation on the average of a sample}
\usage{
influenceOfGivenObservation(
  data,
  weightCol,
  yearCol,
  year0 = NULL,
  year1 = NULL,
  variables,
  indexList,
  ids = NULL,
  idCol = NULL,
  whichRows = NULL,
  reference = NULL,
  method = c("oneYear", "delta", "relativeDelta"),
  missingIn = c("anyYear", "year0", "year1"),
  compareFunc = function(outside, inside) return(inside/outside - 1),
  nCores = 1,
  parallelCluster = NULL,
  verbose = FALSE,
  calledInternally = FALSE
)
}
\arguments{
\item{data}{A data.frame containing all columns specified in \code{weightCol}, \code{yearCol}, \code{indexList}, and \code{variables}.}

\item{weightCol}{The column in \code{data} that contains the weights.}

\item{yearCol}{The column in \code{data} that contains the year.}

\item{year0}{Optional: The reference year to which the delta should be calculated. Only has to be given if \code{method = "oneYear"} is chosen.}

\item{year1}{Optional: The subsequent year. Only has to be given if \code{method = "oneYear"} is chosen.}

\item{variables}{The variables for which the calculation should be executed. Consult the example in \code{\link{findInfluencialObservations}} to see how this argument must be structured.}

\item{indexList}{A list which holds all index levels which should be checked. Consult the example in \code{\link{findInfluencialObservations}} to see how this argument must be structured.}

\item{ids}{Optional: The ids of the observations of which the influence should be calculated. If \code{NULL}, then \code{whichRows} has to be specified.}

\item{idCol}{Optional: The name of the column in \code{data} which contains the ids. Must be specified if \code{ids} is not \code{NULL}.}

\item{whichRows}{Optional: The row numbers of the observations of which the influence should be calculated. If \code{NULL}, then \code{ids} has to be specified.}

\item{reference}{Optional: Reference average values for year0 against which the comparison should be made. If \code{NULL}, then the average values of \code{year0} will be calculated inside the function.}

\item{method}{Either \code{"oneYear"}, \code{"delta"} or \code{"relativeDelta"}, specifying if the influence on one year or on the absolulte/relative delta between two years should be calculated.}

\item{compareFunc}{The function to compare averages when the observation is part of the sample (\code{inside}) or outside the sample (\code{outside}).}

\item{nCores}{The number of cores to use for the computations.}

\item{parallelCluster}{Optional parallel cluster created with \code{\link[snow:makeCluster]{snow::makeCluster}} and registered with \code{\link[doSNOW:registerDoSNOW]{doSNOW::registerDoSNOW}}.}
}
\value{
A list containing the results for each individual observation that was checked.
}
\description{
Calculate the influence of a given observation on the average of a sample
}
\examples{
# See \link{findInfluencialObservations}

}
\author{
Daniel Hoop
}
