% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multiCorePost}
\alias{multiCorePost}
\title{Multicore post operations}
\usage{
multiCorePost(settings)
}
\arguments{
\item{settings}{The return value of \code{\link{multiCorePrep}}.}
}
\description{
Stops the parallel cluster if it was initialized.
}
\author{
Daniel Hoop
}
\keyword{internal}
