% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multiCorePrep}
\alias{multiCorePrep}
\title{Multicore preparations}
\usage{
multiCorePrep(nCores, parallelCluster)
}
\arguments{
\item{nCores}{The number of cores to register.}

\item{parallelCluster}{Optionally: A parallel cluster that was already registered.}
}
\value{
A list with names \code{"clusterWasRegistered"} and \code{"parallelCluster"}.
}
\description{
Installs required packages, registers a parallel cluster and returns it as value, if on Windows operating systems.
}
\author{
Daniel Hoop
}
\keyword{internal}
